+++ NEWS.md
@@ -8,6 +8,7 @@ None
 2. Fix #74 - Add function `make_classification_base_tbl()`
 3. Fix #77 - Add function `internal_make_spec_tbl()`
 4. Fix #78 - Add function `internal_set_args_to_tune()`
+5. Fix #16 - Add function `create_workflow_set()`
 
 ## Minor Fixes and Improvements
 1. Fix #72 - Update `fast_classification_parsnip_spec_tbl()` and